requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
aiosqlite>=0.18.0
APScheduler>=3.10.4
aiohttp>=3.8.5
pydantic>=1.10.0
loguru>=0.7.0
PyYAML>=6.0
